hero: move by fractional pixel increments

+++ engine/include/rage1/hero.h
@@ -50,7 +50,7 @@ struct hero_animation_data_s {
 #define MOVE_ALL	( MOVE_UP | MOVE_DOWN | MOVE_LEFT | MOVE_RIGHT )
 struct hero_movement_data_s {
     uint8_t last_direction;
-    uint8_t dx,dy;
+    uint16_t dx,dy;
 };
 
 // describes the damage mode for the hero
